% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{food}
\alias{food}
\title{Food consumption figures for countries in Europe}
\value{
A tibble or data frame.
}
\description{
Taken from a larger data set of the world
}
\examples{
data(food)
}
\keyword{datasets}
